Transformed from uint to size_t and got rid of more warnings

+++ multinet/include/community.h
@@ -12,7 +12,7 @@ struct layer_set_comparator {
 		if (a.size() != b.size()) return a.size() < b.size();
 		sorted_set<LayerSharedPtr>::iterator it1 = a.begin();
 		sorted_set<LayerSharedPtr>::iterator it2 = b.begin();
-		for (uint i=0; i<a.size(); i++) {
+		for (size_t i=0; i<a.size(); i++) {
 			if ((*it1)<(*it2))
 				return true;
 			if ((*it1)>(*it2))
@@ -25,11 +25,12 @@ struct layer_set_comparator {
 typedef std::set<sorted_set<LayerSharedPtr>, layer_set_comparator > layer_sets;
 
 class cpm_community {
+
 	public:
 		cpm_community(long id, hash_set<CliqueSharedPtr> cliques, hash_set<LayerSharedPtr> layers);
 		cpm_community();
 		std::set<ActorSharedPtr> actors();
-		int size() const;
+		size_t size() const;
 		bool operator==(const cpm_community& comp) const;
 		bool operator!=(const cpm_community& comp) const;
 		bool operator<(const cpm_community& comp) const;
@@ -43,29 +44,29 @@ class cpm_community {
 
 	typedef std::shared_ptr<cpm_community> CommunitySharedPtr;
 
-	hash_set<CommunitySharedPtr> ml_cpm(MLNetworkSharedPtr mnet, uint k, uint m1, int m2, int m3);
+	hash_set<CommunitySharedPtr> ml_cpm(MLNetworkSharedPtr mnet, size_t k, size_t m1, int m2, int m3);
 
 	hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> > get_mlneighbors(MLNetworkSharedPtr mnet, ActorSharedPtr actor);
 
 	hash_set<LayerSharedPtr> neighboring_layers(MLNetworkSharedPtr mnet, ActorSharedPtr actor1, ActorSharedPtr actor2);
 
-	hash_set<CliqueSharedPtr> find_max_cliques(MLNetworkSharedPtr mnet, hash_set<LayerSharedPtr> layers, uint k, uint m);
+	hash_set<CliqueSharedPtr> find_max_cliques(MLNetworkSharedPtr mnet, hash_set<LayerSharedPtr> layers, size_t k, size_t m);
 
-	hash_set<CliqueSharedPtr> find_max_cliques(MLNetworkSharedPtr mnet, uint k, uint m);
-	hash_set<CliqueSharedPtr> find_max_cliques_it(MLNetworkSharedPtr mnet, uint k, uint m);
+	hash_set<CliqueSharedPtr> find_max_cliques(MLNetworkSharedPtr mnet, size_t k, size_t m);
+	hash_set<CliqueSharedPtr> find_max_cliques_it(MLNetworkSharedPtr mnet, size_t k, size_t m);
 
 	void find_max_cliques(MLNetworkSharedPtr mnet, CliqueSharedPtr& A, hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> >& B, hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> >& C,
-			hash_set<ActorSharedPtr>& in_clique, hash_set<CliqueSharedPtr>& result, uint k, uint m);
-	void find_max_cliques_it(MLNetworkSharedPtr mnet, CliqueSharedPtr& A, hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> >& B, hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> >& C, hash_set<CliqueSharedPtr>& result, uint k, uint m);
+			hash_set<ActorSharedPtr>& in_clique, hash_set<CliqueSharedPtr>& result, size_t k, size_t m);
+	void find_max_cliques_it(MLNetworkSharedPtr mnet, CliqueSharedPtr& A, hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> >& B, hash_map<ActorSharedPtr,hash_set<LayerSharedPtr> >& C, hash_set<CliqueSharedPtr>& result, size_t k, size_t m);
 
 	std::map<CliqueSharedPtr,hash_set<CliqueSharedPtr> > build_adjacency_graph(const hash_set<CliqueSharedPtr>& C, int m);
 
 	std::map<CliqueSharedPtr,hash_set<CliqueSharedPtr> > build_max_adjacency_graph(const hash_set<CliqueSharedPtr>& C, int k, int m);
 
-	hash_set<CommunitySharedPtr> find_max_communities_max_layers(MLNetworkSharedPtr mnet, const clique_adjacency_graph& adjacency, uint m);
+	hash_set<CommunitySharedPtr> find_max_communities_max_layers(MLNetworkSharedPtr mnet, const clique_adjacency_graph& adjacency, size_t m);
 
 	void find_max_communities_max_layers(const clique_adjacency_graph& adjacency, CommunitySharedPtr& A,
-			vector<CliqueSharedPtr> Candidates, hash_set<CliqueSharedPtr>& processedCliques, layer_sets& processedLayerCombinations, uint m, hash_set<CommunitySharedPtr>& result);
+			vector<CliqueSharedPtr> Candidates, hash_set<CliqueSharedPtr>& processedCliques, layer_sets& processedLayerCombinations, size_t m, hash_set<CommunitySharedPtr>& result);
 };
 
 #endif /* MULTIPLENETWORK_COMMUNITY_H_ */
